Add devise for user sign-in management
We will need user accounts to assign questions to known users.

I've added the devise gem to handle the process of registration and
sign-in, because it allows me to bootstrap quickly. This commit contains
the result of running `rails generate devise:install`, and following the
additional instructions of adding a `root` route (which I've pointed to
a QuestionsController that does nothing at the moment), and adding flash
messages to the layout.

For a real app, I would spend more time ensuring the correct
registration flow exists, but for this app right now the default is
sufficient.