Require login for every page view
This commit ensures that all pages require a user account. This is the
result of running `rails g devise User` to create a User record set up
for Devise. I then added the before_action in ApplicationController.

This is the simplest way to get login into the workflow for now. In
future, it might make sense to allow non-logged in users (and therefore
web crawlers) to view questions and answers without an account, but that
can be done later.