Add name to User
We will need a name of the user to display next to questions, and email
isn't appropriate because most users would like to keep that kind of
info private.

This commit adds a name column to user and adds it to the registration
form. I've decided to make this a unique nickname so users are easily
identifiable on the platform.

Choosing a non-unique name will cause errors, so in future it would be
good to make this user flow better, and maybe suggest names for users.