Allow user to create new question
Questions are the core of this site. They have a title and a body and
are linked to an author, which is a User account.

This commit contains the code to create and view questions. It's a
standard set of resourceful Rails routes and controller actions,
referring to a new Question model.

At present, the Question model has some standard validations which
aren't interesting enough to test.

There is some duplication in the feature code, specifically around
managing ActiveRecord objects. I'm reasonably comfortable with this at
this point; if we introduce another feature that manages users and login
then it'd be good to move the code into a helper.

I've resisted introducing FactoryBot to manage fixtures just for
simplicity, but as we create more objects this will probably be
required.