Allow users to answer questions
This commit introduces answers, so questions can be answered. At the
bottom of the question page there is now a list of answers and a form to
add a new answer.

A question can have many answers, including many by the same user.
Answers are displayed in creation order. I added an index to
`created_at` to make this more efficient. With pristine data, it would
be possible to rely on the primary key to derive the order. However,
there are scenarios like importing other datasets that might mean
insertion id is not sufficient and I'd rather not build in the
assumption that id is enough.

For time reasons, I haven't added as many unit specs, relying solely on
the feature spec. There is also some more duplication around creation of
users in the spec which would be good do remove soon.